feat: modify type of Post
Id (from Integer to Long) IsBest(from Integer to Boolean) IsDeleted(from Integer to Boolean)

+++ homepage/src/main/java/com/example/homepage/PostDir/post/service/PostService.java
@@ -7,8 +7,8 @@ public interface PostService {
     PostResponseListDTO getHome();
     PostResponseListDTO getCursor(int cursor, int limit);
     void createPost(PostRequestDTO postRequestDTO);
-    void updatePost(Integer id, PostRequestDTO postRequestDTO);
-    void deletePost(Integer id);
-    void setPostBest(Integer id, String username, String password);
-    void cancelPostBest(Integer id, String username, String password);
+    void updatePost(Long id, PostRequestDTO postRequestDTO);
+    void deletePost(Long id);
+    void setPostBest(Long id, String username, String password);
+    void cancelPostBest(Long id, String username, String password);
 }
